+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-#Changelog React Log Lifecycle
+# Changelog React Log Lifecycle
+
+## 0.0.2
+
+* Add optional object logging with a flags object
 
 ## 0.0.1